Teste revert

+++ ApiPreco/SensorReadDataApi/Domain/Price.cs
@@ -8,7 +8,7 @@ namespace PrecoApi.Domain
     public class Price
     {
         public long ProductId { get; set; }
-        public float SalePrice { get; set; }
+        public decimal SalePrice { get; set; }
         public decimal MaximumPrice { get; set; }
         public decimal PercentageDiscount { get; set; }
         public string DiscountType { get; set; }
